Error prevention
When a player uses nitrous it tries to sync the flames to all clients. But if the player is out of the scope the users client they will not be able to resolve the entity attached to the network id. Just a simple check if the client was able to resolve an entity from the network id